pyproject.toml
[project]
name = "simplepyagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
]
